pyproject.toml
[project]
name = "aoc25"
version = "0.1.0"
description = "Solutions to Advent of Code 2025 in Python"
readme = "README.md"
authors = [
    { name = "Daniel Holmes", email = "danieljholmes@hotmail.co.uk" }
]
requires-python = ">=3.13"
dependencies = [
    "aoc-tools-dannyboywoop>=2.0.0",
]

[project.scripts]
aoc25 = "aoc25:main"

[build-system]
requires = ["uv_build>=0.8.4,<0.9.0"]
build-backend = "uv_build"

[dependency-groups]
dev = [
    "mypy>=1.19.0",
    "pytest>=9.0.1",
    "ruff>=0.14.7",
]
